backend/requirements.txt
# Python dependencies for Calm Table backend.
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.4,<5.0
django-filter>=24.2,<25.0
djangorestframework-simplejwt>=5.3,<6.0
drf-spectacular>=0.27,<1.0
django-environ>=0.11,<1.0
psycopg2-binary>=2.9,<3.0
celery>=5.4,<6.0
redis>=5.0,<6.0
gunicorn>=21.2,<22.0
flake8>=7.0,<8.0
pytest>=8.2,<9.0
pytest-django>=4.8,<5.0
pytest-cov>=5.0,<6.0
